src/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
gevent==21.12.0
pymysql==1.0.2
pyyaml==5.3.1
scp==0.14.5
paramiko==2.7.2
requests==2.28.1
six==1.16.0
docker==6.0.1
